# openapi.yaml
openapi: 3.0.3
info:
  title: Lymchic Store API
  version: 1.0.0
  description: API RESTful para la tienda Lymchic Store
servers:
  - url: http://localhost:3000/api
paths:
...
  /products:
    get:
      summary: Listar productos
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [name, price, description, id, image, size, color, featured, categoryId] }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc] }
        - in: query
          name: category_id
          schema: { type: string, format: uuid }
        - in: query
          name: min_price
          schema: { type: number }
        - in: query
          name: max_price
          schema: { type: number }
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              example:
                data:
                  - id: "b1e2c3d4-5678-1234-9abc-1234567890ab"
                    name: Camiseta
                    price: 25.5
                    description: Camiseta de algodón
                    image: https://img.com/camisa.png
                    size: M
                    color: rojo
                    featured: true
                    categoryId: "a1b2c3d4-5678-1234-9abc-1234567890ab"
                pagination:
                  total: 1
                  page: 1
                  limit: 10
                  totalPages: 1
  /products/{id}:
    get:
      summary: Obtener producto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              example:
                id: "b1e2c3d4-5678-1234-9abc-1234567890ab"
                name: Camiseta
                price: 25.5
                description: Camiseta de algodón
                image: https://img.com/camisa.png
                size: M
                color: rojo
                featured: true
                categoryId: "a1b2c3d4-5678-1234-9abc-1234567890ab"
        '400':
          description: ID inválido
          content:
            application/json:
              example:
                message: Invalid product ID
        '404':
          description: Producto no encontrado
          content:
            application/json:
              example:
                message: Product not found
...